Consolidate resolveCanvasFontFamily into canvasRenderHelpers
Remove the single-function canvasTypography.ts module and move resolveCanvasFontFamily into canvasRenderHelpers.ts. Update all import sites accordingly.
